requirements.txt
pyserial >= 3.1.1